Fix bad argument when going out of Wasm

+++ core/shared/platform/common/posix/posix_scheduler.c
@@ -26,7 +26,7 @@ os_sched_getscheduler(uint pid)
 }
 
 int os_sched_setaffinity(uint pid,
-                         size_t cpu_set_size,
+                         uint cpu_set_size,
                          const void *mask)
 {
     return sched_setaffinity(pid, cpu_set_size, mask);
